requirement.txt
asgiref==3.2.7
Django==3.0.5
django-adaptors==0.2.5
django-admin-interface==0.12.1
django-colorfield==0.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-widget-tweaks==1.4.8
html5lib==1.0.1
Jinja2==2.11.1
Pillow==7.1.2
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
